Keep interface clear
Encapsulation the struct and keep interface clear.
